pyproject.toml
[tool.poetry]
name = "overlay-analytics"
version = "0.1.0"
description = ""
authors = ["deepsp94 <anantsp94@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
eth-brownie = "^1.19.1"
jupyter = "^1.0.0"
pandas = "^1.5.0"
matplotlib = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
